+++ Patches/PlayerControllerB_Patches.cs
@@ -14,7 +14,7 @@ internal class PlayerControllerB_Patches
             if (SyncedConfig.Instance.EnemiesData.ContainsKey(enemy.enemyType.enemyName))
                 continue;
 
-            EnemiesDataManager.RegisterEnemy(enemy.enemyType.enemyName, new(true, Metadata: new(new(0, 0, 0), new(0, 0, 0), true)));
+            EnemiesDataManager.RegisterEnemy(enemy.enemyType.enemyName, new(true, metadata: new(new(0, 0, 0), new(0, 0, 0), new(0, 0, 0), true)));
             Plugin.logger.LogInfo($"Mob was not registered. Registered it with name '{enemy.enemyType.enemyName}'");
         }
         EnemiesDataManager.SaveEnemiesData();